add localization for reinforced universal implement
